Implemented device code, in a slightly different way, the user would login, and request and a device code from their phone, computer, then enter the code on their TV, Fridge, etc to login. This implementation eliminates the need for polling the token endpoint, and simplifies the process. Refactor seed data using partial static classes for better Single Responsibility.
